Move root command to cmds/ subfolder
