+++ cmds/config/config.go
@@ -7,8 +7,8 @@ import (
 	"os"
 
 	"github.com/spf13/cobra"
+	"github.com/taskcluster/taskcluster-cli/cmds/root"
 	"github.com/taskcluster/taskcluster-cli/config"
-	"github.com/taskcluster/taskcluster-cli/root"
 	"github.com/taskcluster/taskcluster-client-go"
 )
