+++ cmds/task/task.go
@@ -1,7 +1,7 @@
 package task
 
 import (
-	"github.com/taskcluster/taskcluster-cli/root"
+	"github.com/taskcluster/taskcluster-cli/cmds/root"
 
 	"github.com/spf13/cobra"
 )
